fix compilation errors for `default-features = false`
